Can encode data into qr code standards now

+++ src/constants.c
@@ -1,6 +1,6 @@
 #include "constants.h"
 
 // These all follow the versions 1-9, 10-26, 27-40
-const int NUMERIC_CHARACTER_COUNT_LEN[3] = {10, 12, 14};
-const int ALPHANUMERIC_CHARACTER_COUNT_LEN[3] = {9, 11, 13};
-const int BYTE_CHARACTER_COUNT_LEN[3] = {8, 16, 16};
+const uint8_t NUMERIC_CHARACTER_COUNT_LEN[3] = {10, 12, 14};
+const uint8_t ALPHANUMERIC_CHARACTER_COUNT_LEN[3] = {9, 11, 13};
+const uint8_t BYTE_CHARACTER_COUNT_LEN[3] = {8, 16, 16};
